Remove activity-log/display-rules and jetpack/only-realtime-products flags
