--- client/components/jetpack/card/product-without-price/style.scss
@@ -1,121 +1,0 @@
-@import '@wordpress/base-styles/breakpoints';
-@import '@wordpress/base-styles/mixins';
-
-.product-without-price,
-.product-without-price--full-width {
-	padding: 24px 32px;
-}
-
-.product-without-price--full-width {
-	@include break-small {
-		display: flex;
-		flex-direction: column;
-	}
-
-	@include breakpoint-deprecated( '>660px' ) {
-		flex-direction: row;
-	}
-
-	.product-without-price__header {
-		@include breakpoint-deprecated( '>660px' ) {
-			flex: 1;
-			margin-inline-end: 32px;
-		}
-	}
-
-	.product-without-price__title {
-		@include breakpoint-deprecated( '>660px' ) {
-			margin-bottom: initial;
-		}
-	}
-
-	.product-without-price__button.button {
-		@include breakpoint-deprecated( '>660px' ) {
-			max-width: 210px;
-		}
-
-		@include break-large {
-			margin-bottom: 0;
-		}
-	}
-
-	.product-without-price__features-list {
-		flex: 1;
-
-		@include break-large {
-			align-self: center;
-
-			display: flex;
-			align-items: center;
-			flex-wrap: wrap;
-			flex: 2 1;
-
-			margin-top: initial;
-			margin-bottom: -14px;
-
-			> li {
-				width: calc( 50% - 8px );
-				padding-inline-start: 8px;
-
-				margin-bottom: 14px;
-			}
-
-			> li:last-child {
-				margin-bottom: 14px;
-			}
-		}
-	}
-}
-
-.product-without-price__title {
-	color: var( --studio-gray-100 );
-	margin: 12px 0 20px;
-
-	font-family: 'Noto Sans SemiCondensed Bold', 'Noto Sans', $sans;
-	font-size: 1.875rem;
-	font-weight: 700;
-	line-height: inherit;
-
-	@include breakpoint-deprecated( '>660px' ) {
-		margin-top: initial;
-	}
-}
-
-.product-without-price__subheadline {
-	// 2 lines, accounting for line-height of 1.5
-	min-height: calc( 2rem * 1.5 );
-
-	color: var( --studio-gray-60 );
-	font-size: inherit;
-}
-
-.product-without-price__button.button {
-	display: block;
-
-	width: 100%;
-	margin: 24px 0;
-
-	color: var( --studio-jetpack-green-40 );
-	border: solid 2px currentColor;
-	border-radius: 4px;
-
-	font-size: 1rem;
-}
-
-.product-without-price__features-list {
-	color: var( --studio-gray-60 );
-	margin: 0 16px 32px;
-
-	> li {
-		margin-bottom: 16px;
-		padding-inline-start: 8px;
-	}
-
-	> li:last-child {
-		margin-bottom: initial;
-	}
-
-	@include breakpoint-deprecated( '>660px' ) {
-		margin: 16px;
-	}
-}
