--- client/my-sites/plans/jetpack-plans/jetpack-crm-free-card/style.scss
@@ -1,72 +1,0 @@
-.jetpack-plans__iteration--only-realtime-products {
-	.jetpack-free-card .display-price__price-free {
-		padding-block-start: 24px;
-	}
-
-	.jetpack-crm-free-card,
-	.jetpack-free-card {
-		.display-price {
-			min-height: 133px;
-			margin-block-start: 38px;
-		}
-
-		.display-price__above-price-text {
-			display: block;
-			margin-inline-start: 2px;
-
-			font-size: 1rem;
-			font-weight: 700;
-
-			// Adjust leading margin so that different UI fonts
-			// still align to the price the same way
-			&.is-jetpack-cloud {
-				margin-inline-start: 4px;
-			}
-		}
-
-		.display-price__price-free {
-			display: block;
-			line-height: $font-headline-medium;
-			letter-spacing: -2px;
-		}
-	}
-}
-
-.jetpack-crm-free-card.product-without-price--full-width {
-	display: flex;
-	column-gap: 64px;
-	flex-wrap: wrap;
-	flex-grow: 1;
-
-	.product-without-price__header {
-		@include breakpoint-deprecated( '>660px' ) {
-			flex: 1 0 calc( 300px - 64px - 1px );
-			margin-inline-end: initial;
-		}
-	}
-
-	.product-without-price__features-list {
-		@include breakpoint-deprecated( '>660px' ) {
-			flex: 1 0 calc( 300px - 64px - 1px );
-			align-self: center;
-			margin: 0 0 0 16px;
-		}
-
-		display: block;
-		margin: 0 16px 32px;
-
-		& > li {
-			width: initial;
-
-			&:last-child {
-				margin-block-end: initial;
-			}
-		}
-	}
-
-	.product-without-price__button.button {
-		@include breakpoint-deprecated( '>660px' ) {
-			max-width: 210px;
-		}
-	}
-}
