+++ client/my-sites/plans/jetpack-plans/jetpack-free-card/index.tsx
@@ -1,18 +1,58 @@
-import {
-	getForCurrentCROIteration,
-	Iterations,
-} from 'calypso/my-sites/plans/jetpack-plans/iterations';
-import CardWithPrice, { CardWithPriceProps } from './with-price';
-import CardWithoutPrice, { CardWithoutPriceProps } from './without-price';
+import { useTranslate } from 'i18n-calypso';
+import { FC, useMemo } from 'react';
+import JetpackProductCard from 'calypso/components/jetpack/card/jetpack-product-card';
+import useJetpackFreeButtonProps from './use-jetpack-free-button-props';
+import type { QueryArgs } from 'calypso/my-sites/plans/jetpack-plans/types';
 
-const Wrapper: React.FC< CardWithPriceProps | CardWithoutPriceProps > = ( props ) => {
-	return getForCurrentCROIteration( ( iteration: Iterations ) => {
-		if ( iteration === Iterations.ONLY_REALTIME_PRODUCTS ) {
-			return <CardWithPrice { ...props } />;
-		}
-
-		return <CardWithoutPrice { ...props } />;
-	} );
+type OwnProps = {
+	siteId: number | null;
+	urlQueryArgs: QueryArgs;
 };
 
-export default Wrapper;
+const useFreeItem = () => {
+	const translate = useTranslate();
+
+	return useMemo(
+		() => ( {
+			productSlug: 'free',
+			isFree: true,
+			displayName: translate( 'Jetpack Free' ),
+			features: {
+				items: [
+					{ text: translate( 'Brute force attack protection' ) },
+					{ text: translate( 'Site stats' ) },
+					{ text: translate( 'Content Delivery Network' ) },
+				],
+			},
+		} ),
+		[ translate ]
+	);
+};
+
+const JetpackFreeCard: FC< OwnProps > = ( { siteId, urlQueryArgs } ) => {
+	const translate = useTranslate();
+	const freeProduct = useFreeItem();
+
+	const { href: buttonHref, onClick: onButtonClick } = useJetpackFreeButtonProps(
+		siteId,
+		urlQueryArgs
+	);
+
+	return (
+		<JetpackProductCard
+			className="jetpack-free-card"
+			hideSavingLabel
+			buttonPrimary
+			item={ freeProduct }
+			headerLevel={ 3 }
+			description={ translate(
+				'Power up your WordPress site with essential security and performance features.'
+			) }
+			buttonLabel={ translate( 'Start for free' ) }
+			buttonURL={ buttonHref }
+			onButtonClick={ onButtonClick }
+		/>
+	);
+};
+
+export default JetpackFreeCard;
