+++ client/my-sites/plans/jetpack-plans/product-grid/style.scss
@@ -67,8 +67,8 @@
 		height: 100%;
 	}
 
-	&.second-grid {
-		margin: 16px 0;
+	.jetpack-product-card.jetpack-free-card {
+		margin: initial;
 	}
 }
 
@@ -81,103 +81,6 @@
 	}
 }
 
-.product-grid__free {
-	height: 100%;
-	display: flex;
-	gap: 16px;
-	flex-wrap: wrap;
-
-	.product-without-price {
-		flex: 1 1 0;
-		min-width: 300px;
-	}
-
-	&.add-top-margin {
-		margin-block-start: 16px;
-	}
-}
-
 .product-grid__more.is-detached {
 	margin-top: 48px;
-}
-
-// Show both CRM and Free as full-width on narrow or mid-width screens
-.jetpack-plans__iteration--only-realtime-products {
-	.product-grid__free {
-		margin-top: 16px;
-		flex-direction: column;
-
-		@include breakpoint-deprecated( '>660px' ) {
-			flex-direction: initial;
-		}
-	}
-
-	.jetpack-product-card.jetpack-free-card {
-		margin: initial;
-	}
-
-	.product-grid__free .jetpack-crm-free-card .jetpack-product-card__body {
-		.display-price {
-			min-height: 133px;
-		}
-
-		@include breakpoint-deprecated( '>660px' ) {
-			display: grid;
-			grid-column-gap: 32px;
-			grid-template-columns: 1fr 1fr;
-
-			padding-block-end: 32px;
-
-			.jetpack-product-card__product-name {
-				grid-column: 1 / span 2;
-				grid-row: 1;
-			}
-
-			.display-price {
-				grid-column: 1;
-				grid-row: 2;
-
-				min-height: initial;
-			}
-
-			.display-price__above-price-text {
-				margin-block-start: initial;
-			}
-
-			.jetpack-product-card__button {
-				grid-column: 1;
-				grid-row: 3;
-				align-self: flex-end;
-				margin-block-end: initial;
-			}
-
-			.jetpack-product-card__description {
-				grid-column: 2;
-				grid-row: 2;
-				min-height: initial;
-			}
-
-			.jetpack-product-card__features {
-				margin-block-end: initial;
-			}
-
-			.jetpack-product-card__features-list {
-				grid-column: 2;
-				grid-row: 3;
-
-				margin-block-end: initial;
-				margin-inline-end: initial;
-			}
-
-			.jetpack-product-card__features-item {
-				&:first-child {
-					margin-block-start: initial;
-				}
-
-				&:last-child {
-					margin-block-end: initial;
-				}
-			}
-		}
-	}
 }
