+++ client/my-sites/plans/jetpack-plans/use-get-plans-grid-products.ts
@@ -1,9 +1,5 @@
 import {
 	JETPACK_ANTI_SPAM_PRODUCTS,
-	PRODUCT_JETPACK_BACKUP_DAILY,
-	PRODUCT_JETPACK_BACKUP_DAILY_MONTHLY,
-	PRODUCT_JETPACK_BACKUP_REALTIME,
-	PRODUCT_JETPACK_BACKUP_REALTIME_MONTHLY,
 	PRODUCT_JETPACK_BACKUP_T1_YEARLY,
 	PRODUCT_JETPACK_BACKUP_T1_MONTHLY,
 	PRODUCT_JETPACK_BACKUP_T2_YEARLY,
@@ -17,10 +13,6 @@ import {
 	getPlan,
 } from '@automattic/calypso-products';
 import { useSelector } from 'react-redux';
-import {
-	doForCurrentCROIteration,
-	Iterations,
-} from 'calypso/my-sites/plans/jetpack-plans/iterations';
 import getSitePlan from 'calypso/state/sites/selectors/get-site-plan';
 import getSiteProducts from 'calypso/state/sites/selectors/get-site-products';
 import slugToSelectorProduct from './slug-to-selector-product';
@@ -53,48 +45,21 @@ const useSelectorPageProducts = ( siteId: number | null ): PlanGridProducts => {
 
 	const backupProductsToShow: string[] = [];
 
-	doForCurrentCROIteration( ( key ) => {
-		if ( Iterations.ONLY_REALTIME_PRODUCTS === key ) {
-			const ownsBackupT1 =
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T1_YEARLY ) ||
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T1_MONTHLY );
-			const ownsBackupT2 =
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T2_YEARLY ) ||
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T2_MONTHLY );
+	const ownsBackupT1 =
+		ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T1_YEARLY ) ||
+		ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T1_MONTHLY );
+	const ownsBackupT2 =
+		ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T2_YEARLY ) ||
+		ownedProducts.includes( PRODUCT_JETPACK_BACKUP_T2_MONTHLY );
 
-			// If neither T1 or T2 backups are owned, then show T1 backups.
-			// Otherwise the one owned will be displayed via purchasedProducts.
-			if ( ! ownsBackupT1 && ! ownsBackupT2 ) {
-				backupProductsToShow.push(
-					PRODUCT_JETPACK_BACKUP_T1_YEARLY,
-					PRODUCT_JETPACK_BACKUP_T1_MONTHLY
-				);
-			}
-		} else {
-			const ownsDaily =
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_DAILY ) ||
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_DAILY_MONTHLY );
-			const ownsRealtime =
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_REALTIME ) ||
-				ownedProducts.includes( PRODUCT_JETPACK_BACKUP_REALTIME_MONTHLY );
-
-			// Show the Backup product the site owns, and the one it doesn't own.
-			// In other words, always show both Backup Daily and Backup Real-time.
-			if ( ! ownsDaily ) {
-				backupProductsToShow.push(
-					PRODUCT_JETPACK_BACKUP_DAILY,
-					PRODUCT_JETPACK_BACKUP_DAILY_MONTHLY
-				);
-			}
-
-			if ( ! ownsRealtime ) {
-				backupProductsToShow.push(
-					PRODUCT_JETPACK_BACKUP_REALTIME,
-					PRODUCT_JETPACK_BACKUP_REALTIME_MONTHLY
-				);
-			}
-		}
-	} );
+	// If neither T1 or T2 backups are owned, then show T1 backups.
+	// Otherwise the one owned will be displayed via purchasedProducts.
+	if ( ! ownsBackupT1 && ! ownsBackupT2 ) {
+		backupProductsToShow.push(
+			PRODUCT_JETPACK_BACKUP_T1_YEARLY,
+			PRODUCT_JETPACK_BACKUP_T1_MONTHLY
+		);
+	}
 
 	availableProducts = [ ...availableProducts, ...backupProductsToShow ];
