+++ client/my-sites/scan/wpcom-scan-upsell.tsx
@@ -16,10 +16,6 @@ import PromoCardCTA from 'calypso/components/promo-section/promo-card/cta';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import { preventWidows } from 'calypso/lib/formatting';
 import useTrackCallback from 'calypso/lib/jetpack/use-track-callback';
-import {
-	getForCurrentCROIteration,
-	Iterations,
-} from 'calypso/my-sites/plans/jetpack-plans/iterations';
 import SidebarNavigation from 'calypso/my-sites/sidebar-navigation';
 import { canCurrentUser } from 'calypso/state/selectors/can-current-user';
 import { getSelectedSiteSlug, getSelectedSiteId } from 'calypso/state/ui/selectors';
@@ -85,16 +81,10 @@ const ScanUpsellBody: FunctionComponent = () => {
 	const translate = useTranslate();
 	const postCheckoutUrl = window.location.pathname + window.location.search;
 
-	const nonAdminNoticeText =
-		getForCurrentCROIteration( {
-			[ Iterations.ONLY_REALTIME_PRODUCTS ]: translate(
-				'Only site administrators can upgrade to access security scanning.'
-			),
-		} ) ?? translate( 'Only site administrators can upgrade to access daily scanning.' );
-	const buttonText =
-		getForCurrentCROIteration( {
-			[ Iterations.ONLY_REALTIME_PRODUCTS ]: translate( 'Get Jetpack Scan' ),
-		} ) ?? translate( 'Get daily scanning' );
+	const nonAdminNoticeText = translate(
+		'Only site administrators can upgrade to access security scanning.'
+	);
+	const buttonText = translate( 'Get Jetpack Scan' );
 
 	return (
 		<PromoCard
