+++ config/horizon.json
@@ -15,7 +15,6 @@
 	"difm_typeform_id": "YMiXMYId",
 	"difm_signup_typeform_id": "m3s2pdgI",
 	"features": {
-		"activity-log/display-rules": true,
 		"async-payments": false,
 		"ad-tracking": false,
 		"calypsoify/plugins": true,
@@ -53,7 +52,6 @@
 		"jetpack/features-section/atomic": true,
 		"jetpack/features-section/jetpack": true,
 		"jetpack/magic-link-signup": true,
-		"jetpack/only-realtime-products": true,
 		"jetpack/simplify-pricing-structure": false,
 		"jetpack/siteless-checkout": true,
 		"jetpack/user-licensing": true,
