+++ config/jetpack-cloud-horizon.json
@@ -24,7 +24,6 @@
 		}
 	],
 	"features": {
-		"activity-log/display-rules": true,
 		"activity-log/v2": true,
 		"always_use_logout_url": false,
 		"current-site/domain-warning": false,
@@ -33,7 +32,6 @@
 		"checkout/google-pay": true,
 		"jetpack-cloud": true,
 		"jetpack/backups-date-picker": true,
-		"jetpack/only-realtime-products": true,
 		"jetpack/search-product": true,
 		"jetpack/pricing-page": true,
 		"jetpack/server-credentials-advanced-flow": true,
