+++ config/jetpack-cloud-production.json
@@ -26,7 +26,6 @@
 	"oauth_client_id": 69041,
 	"features": {
 		"ad-tracking": true,
-		"activity-log/display-rules": true,
 		"activity-log/v2": true,
 		"always_use_logout_url": true,
 		"current-site/domain-warning": false,
@@ -36,7 +35,6 @@
 		"i18n/community-translator": false,
 		"jetpack-cloud": true,
 		"jetpack/backups-date-picker": true,
-		"jetpack/only-realtime-products": true,
 		"jetpack/pricing-page": true,
 		"jetpack/redirect-legacy-plans": true,
 		"jetpack/server-credentials-advanced-flow": true,
