+++ config/jetpack-cloud-stage.json
@@ -26,7 +26,6 @@
 	"oauth_client_id": 69040,
 	"features": {
 		"ad-tracking": false,
-		"activity-log/display-rules": true,
 		"activity-log/v2": true,
 		"always_use_logout_url": true,
 		"current-site/domain-warning": false,
@@ -36,7 +35,6 @@
 		"jetpack-cloud": true,
 		"jetpack/activity-log-sharing": true,
 		"jetpack/backups-date-picker": true,
-		"jetpack/only-realtime-products": true,
 		"jetpack/partner-portal-payment": true,
 		"jetpack/pricing-page": true,
 		"jetpack/redirect-legacy-plans": true,
