+++ config/production.json
@@ -18,7 +18,6 @@
 	"difm_signup_typeform_id": "m3s2pdgI",
 	"features": {
 		"ad-tracking": true,
-		"activity-log/display-rules": true,
 		"async-payments": false,
 		"bilmur-script": true,
 		"calypsoify/plugins": true,
@@ -53,7 +52,6 @@
 		"jetpack/features-section/jetpack": true,
 		"jetpack/features-section/simple": true,
 		"jetpack/magic-link-signup": true,
-		"jetpack/only-realtime-products": true,
 		"jetpack/redirect-legacy-plans": true,
 		"jetpack/simplify-pricing-structure": false,
 		"jetpack/siteless-checkout": true,
