+++ packages/calypso-products/src/plans-list.tsx
@@ -943,73 +943,37 @@ const getPlanJetpackCompleteDetails = (): IncompleteJetpackPlan => ( {
 	availableFor: ( plan ) =>
 		[ PLAN_JETPACK_FREE, ...JETPACK_SECURITY_PLANS, ...JETPACK_LEGACY_PLANS ].includes( plan ),
 	getDescription: () =>
-		isEnabled( 'jetpack/only-realtime-products' )
-			? translate(
-					'Get the full power of Jetpack with all Security, Performance, Growth, and Design tools.'
-			  )
-			: translate(
-					'Get the full Jetpack suite with real-time security, enhanced search, CRM, marketing, growth, and design tools.'
-			  ),
+		translate(
+			'Get the full power of Jetpack with all Security, Performance, Growth, and Design tools.'
+		),
 	getTagline: () => translate( 'For best-in-class WordPress sites' ),
-	getPlanCardFeatures: () =>
-		isEnabled( 'jetpack/only-realtime-products' )
-			? [
-					FEATURE_JETPACK_ALL_BACKUP_SECURITY_FEATURES,
-					FEATURE_JETPACK_PRODUCT_VIDEOPRESS,
-					FEATURE_PRODUCT_SEARCH_V2,
-					FEATURE_CRM_V2,
-			  ]
-			: [
-					FEATURE_PLAN_SECURITY_DAILY,
-					FEATURE_BACKUP_REALTIME_V2,
-					FEATURE_PRODUCT_SCAN_REALTIME_V2,
-					FEATURE_CRM_V2,
-					FEATURE_PRODUCT_SEARCH_V2,
-			  ],
-	getIncludedFeatures: () =>
-		isEnabled( 'jetpack/only-realtime-products' )
-			? [
-					FEATURE_JETPACK_BACKUP_T2_YEARLY,
-					FEATURE_JETPACK_BACKUP_T2_MONTHLY,
-					FEATURE_JETPACK_SCAN_DAILY,
-					FEATURE_JETPACK_SCAN_DAILY_MONTHLY,
-					FEATURE_JETPACK_ANTI_SPAM,
-					FEATURE_JETPACK_ANTI_SPAM_MONTHLY,
-					FEATURE_JETPACK_SEARCH,
-					FEATURE_JETPACK_SEARCH_MONTHLY,
-					FEATURE_JETPACK_CRM,
-					FEATURE_JETPACK_CRM_MONTHLY,
-					FEATURE_BACKUP_ARCHIVE_UNLIMITED,
-					FEATURE_VIDEO_UPLOADS_JETPACK_PRO,
-					FEATURE_REPUBLICIZE,
-					FEATURE_ADVANCED_SEO,
-					FEATURE_SEO_PREVIEW_TOOLS,
-					FEATURE_SIMPLE_PAYMENTS,
-					FEATURE_WORDADS_INSTANT,
-					FEATURE_GOOGLE_ANALYTICS,
-					FEATURE_PREMIUM_SUPPORT,
-			  ]
-			: [
-					FEATURE_JETPACK_BACKUP_REALTIME,
-					FEATURE_JETPACK_BACKUP_REALTIME_MONTHLY,
-					FEATURE_JETPACK_SCAN_DAILY,
-					FEATURE_JETPACK_SCAN_DAILY_MONTHLY,
-					FEATURE_JETPACK_ANTI_SPAM,
-					FEATURE_JETPACK_ANTI_SPAM_MONTHLY,
-					FEATURE_JETPACK_SEARCH,
-					FEATURE_JETPACK_SEARCH_MONTHLY,
-					FEATURE_JETPACK_CRM,
-					FEATURE_JETPACK_CRM_MONTHLY,
-					FEATURE_BACKUP_ARCHIVE_UNLIMITED,
-					FEATURE_VIDEO_UPLOADS_JETPACK_PRO,
-					FEATURE_REPUBLICIZE,
-					FEATURE_ADVANCED_SEO,
-					FEATURE_SEO_PREVIEW_TOOLS,
-					FEATURE_SIMPLE_PAYMENTS,
-					FEATURE_WORDADS_INSTANT,
-					FEATURE_GOOGLE_ANALYTICS,
-					FEATURE_PREMIUM_SUPPORT,
-			  ],
+	getPlanCardFeatures: () => [
+		FEATURE_JETPACK_ALL_BACKUP_SECURITY_FEATURES,
+		FEATURE_JETPACK_PRODUCT_VIDEOPRESS,
+		FEATURE_PRODUCT_SEARCH_V2,
+		FEATURE_CRM_V2,
+	],
+	getIncludedFeatures: () => [
+		FEATURE_JETPACK_BACKUP_T2_YEARLY,
+		FEATURE_JETPACK_BACKUP_T2_MONTHLY,
+		FEATURE_JETPACK_SCAN_DAILY,
+		FEATURE_JETPACK_SCAN_DAILY_MONTHLY,
+		FEATURE_JETPACK_ANTI_SPAM,
+		FEATURE_JETPACK_ANTI_SPAM_MONTHLY,
+		FEATURE_JETPACK_SEARCH,
+		FEATURE_JETPACK_SEARCH_MONTHLY,
+		FEATURE_JETPACK_CRM,
+		FEATURE_JETPACK_CRM_MONTHLY,
+		FEATURE_BACKUP_ARCHIVE_UNLIMITED,
+		FEATURE_VIDEO_UPLOADS_JETPACK_PRO,
+		FEATURE_REPUBLICIZE,
+		FEATURE_ADVANCED_SEO,
+		FEATURE_SEO_PREVIEW_TOOLS,
+		FEATURE_SIMPLE_PAYMENTS,
+		FEATURE_WORDADS_INSTANT,
+		FEATURE_GOOGLE_ANALYTICS,
+		FEATURE_PREMIUM_SUPPORT,
+	],
 	getInferiorFeatures: () => [
 		FEATURE_JETPACK_BACKUP_DAILY,
 		FEATURE_JETPACK_BACKUP_DAILY_MONTHLY,
